# api/openapi.yaml
openapi: 3.0.3
info:
  title: Custos API
  description: |
    AI-powered log intelligence platform.
    SDKs capture errors from your applications and Custos explains them in plain language.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/projects/{id}/keys:
    post:
      tags: [Projects]
      summary: Create an API key for a project
      operationId: createAPIKey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
                  example: production-sdk
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKey"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/projects/{id}/keys/{kid}:
    delete:
      tags: [Projects]
      summary: Revoke an API key
      operationId: revokeAPIKey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: kid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: API key revoked
        "404":
          description: Key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Custos-Key
      description: Per-project API key issued from the dashboard or API
  schemas:
...
    APIKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        key:
          type: string
          description: Raw key value — only returned on creation, never again
        created_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: unauthorized
        message:
          type: string
          example: missing or invalid API key